Fixup: Combine thrift_healthcheck_tests into thrift_tests
